convert: Move CombinedSprite definition and add mmber creation method for RawAPIObject.
