+++ openage/convert/dataformat/aoc/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_py_modules(
 	__init__.py
+	combined_sprite.py
 	expected_pointer.py
 	genie_civ.py
 	genie_connection.py
